+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.16.2-0 / Unreleased
   * [Added] ext-filter: matcher-callback for `tree.filterNodes()` may now return 
     `"branch"` and `"skip"`
+  * [Added] ext-filter: new option`nodata` allows to configure a status node for
+    empty results
   * [Fixed] ext-filter: branch mode honors `autoExpand: true`
   * Update to jQuery UI 1.11.4